feat: Replace mma_truss2 with test case
