--- examples/mma_truss2.log
@@ -1,30 +1,0 @@
-2024-08-18 10:10:28,534 - util - INFO - Started
-
-2024-08-18 10:10:28,556 - util - INFO - outvector1 = [ 0.0000  0.0000  0.0000  0.7500  1.0000  0.7500  0.0000]
-2024-08-18 10:10:28,557 - util - INFO - outvector2 = [ 1.0000  1.0000  1.0000]
-
-2024-08-18 10:10:28,641 - util - INFO - outvector1 = [ 1.0000  0.0000  0.0000  0.7956  0.9684  0.7956 -0.0314]
-2024-08-18 10:10:28,643 - util - INFO - outvector2 = [ 0.9034  1.1619  0.9034]
-2024-08-18 10:10:28,644 - util - INFO - kktnorm    = 0.19521410705028036
-
-2024-08-18 10:10:28,735 - util - INFO - outvector1 = [ 2.0000  0.0000  0.0000  0.8444  0.9297  0.8444 -0.0318]
-2024-08-18 10:10:28,737 - util - INFO - outvector2 = [ 0.8171  1.3341  0.8171]
-2024-08-18 10:10:28,738 - util - INFO - kktnorm    = 0.17995507826628795
-
-2024-08-18 10:10:28,844 - util - INFO - outvector1 = [ 3.0000  0.0000  0.0000  0.9018  0.8933  0.9018 -0.0241]
-2024-08-18 10:10:28,845 - util - INFO - outvector2 = [ 0.7370  1.5019  0.7370]
-2024-08-18 10:10:28,846 - util - INFO - kktnorm    = 0.12521171276457907
-
-2024-08-18 10:10:28,939 - util - INFO - outvector1 = [ 4.0000  0.0000  0.0000  0.8890  0.8889  0.8890 -0.0002]
-2024-08-18 10:10:28,941 - util - INFO - outvector2 = [ 0.7498  1.5002  0.7498]
-2024-08-18 10:10:28,942 - util - INFO - kktnorm    = 0.003942382074755963
-
-2024-08-18 10:10:29,022 - util - INFO - outvector1 = [ 5.0000  0.0000  0.0000  0.8889  0.8889  0.8889 -0.0000]
-2024-08-18 10:10:29,023 - util - INFO - outvector2 = [ 0.7500  1.5000  0.7500]
-2024-08-18 10:10:29,024 - util - INFO - kktnorm    = 0.00015105888755352251
-
-2024-08-18 10:10:29,094 - util - INFO - outvector1 = [ 6.0000  0.0000  0.0000  0.8889  0.8889  0.8889 -0.0000]
-2024-08-18 10:10:29,095 - util - INFO - outvector2 = [ 0.7500  1.5000  0.7500]
-2024-08-18 10:10:29,096 - util - INFO - kktnorm    = 4.1029024538547415e-07
-
-2024-08-18 10:10:29,097 - util - INFO - Finished
