+++ test/reference/test_mma_truss_kkt.txt
@@ -1,0 +1,6 @@
+0.19521410705028036
+0.17995507826628832
+0.12521171276457968
+0.003942382074755426
+0.00015105888755462924
+4.1029024537867163e-07
