+++ test/reference/test_mma_truss_vec1.txt
@@ -1,0 +1,6 @@
+0, 0.79558329, 0.96839475, 0.79558329, -0.03135902
+0, 0.844371, 0.92973895, 0.844371, -0.03177998
+0, 0.90177005, 0.89330344, 0.90177005, -0.0241388
+0.00000000e+00, 8.89028023e-01, 8.88888770e-01, 8.89028023e-01, -1.56227200e-04
+0.00000000e+00, 8.88888720e-01, 8.88889145e-01, 8.88888720e-01, -3.86419988e-07
+0.00000000e+00, 8.88888689e-01, 8.88889139e-01, 8.88888689e-01, -3.37372529e-07
