+++ test/reference/test_mma_truss_vec2.txt
@@ -1,0 +1,6 @@
+0.9033675, 1.16190598, 0.9033675
+0.81707856, 1.33406289, 0.81707856
+0.7369803, 1.5019006, 0.7369803
+0.74984347, 1.50015683, 0.74984347
+0.75000026, 1.49999909, 0.75000026
+0.7500003, 1.49999907, 0.7500003
